Search and get data from Robonomist Data Server

data() is a convenience function that searches the database and returns

search results if multiple matching tables are found, and

data if only one match is found.

data_get() returns data without performing searching or pattern filters. It will fail, if no match exists.

data_search() only performs a search and returns the results.

data(
  pattern = "",
  dl_filter = NULL,
  labels = TRUE,
  lang = NULL,
  na.rm = FALSE,
  ...
)

data_get(id, dl_filter = NULL, labels = TRUE, lang = NULL, na.rm = FALSE, ...)

data_search(pattern = "")

Arguments

pattern

string, Search query or table id, possibly followed by a -filter.

dl_filter

list or named vector passed to datasource download functions for filtering incoming data. Supported by tulli, OECD and ECB.

labels

logical, Some datasources can return labelled or coded data.

lang

Two-letter language code, e.g. "en" or "sv".

na.rm

Px-file based datasources return a table with a combination of all categories. Missing values can be filtered when reading the file to improve preformance.

...

Datasource-specific arguments. TODO

id

string, Exact table id

Details

The data() function allows for a special -filter. When the pattern matches a single table and the function returns a data frame, the -filter can be used to subset rows of data frame. The tibble is filterd by sequence of regular expressions separated by section sign . The regular expression are applied to data frame's variables sequentially.

If the last variable is a date, it is used as a start date filter.